implementados botões de adicionar produtos e ingredientes

+++ GuilhermeBarth-LeandroNeto-VitorHugo/src/Interfaces/TelaAddIngrediente.java
@@ -144,7 +144,7 @@ public class TelaAddIngrediente extends javax.swing.JFrame {
             }
             Ingrediente igrediente = new Ingrediente();
             igrediente.setNome(tf_Nome.getText());
-            igrediente.setCustoUnitario(Float.parseFloat(tf_Valor.getText()));
+            igrediente.setCustoUnitario(Double.parseDouble(tf_Valor.getText()));
             ingredientes.add(igrediente);
             JOptionPane.showMessageDialog(this, "Igrediente cadastrado com sucesso", "Aviso", JOptionPane.INFORMATION_MESSAGE);
             InterfacePrincipal principal = new InterfacePrincipal(pedidos, ingredientes, bebidas, pizzas, clientes);
